+++ com.unity.shadergraph/Editor/Drawing/Inspector/PropertyDrawers/SampleTexture3DNodePropertyDrawer.cs
@@ -10,17 +10,13 @@ using UnityEngine.UIElements;
 
 namespace UnityEditor.ShaderGraph.Drawing.Inspector.PropertyDrawers
 {
-    [SGPropertyDrawer(typeof(SampleTexture2DNode))]
-    class SampleTexture2DNodePropertyDrawer : AbstractMaterialNodePropertyDrawer
+    [SGPropertyDrawer(typeof(SampleTexture3DNode))]
+    class SampleTexture3DNodePropertyDrawer : AbstractMaterialNodePropertyDrawer
     {
         internal override void AddCustomNodeProperties(VisualElement parentElement, AbstractMaterialNode nodeBase, Action setNodesAsDirtyCallback, Action updateNodeViewsCallback)
         {
-            var node = nodeBase as SampleTexture2DNode;
-            PropertyDrawerUtils.AddCustomCheckboxProperty(
-                parentElement, nodeBase, setNodesAsDirtyCallback, updateNodeViewsCallback,
-                "Use Global Mip Bias", "Change Enable Global Mip Bias",
-                () => node.enableGlobalMipBias, (val) => node.enableGlobalMipBias = val);
-            PropertyDrawerUtils.AddCustomEnumProperty<Texture2DMipSamplingMode>(
+            var node = nodeBase as SampleTexture3DNode;
+            PropertyDrawerUtils.AddCustomEnumProperty<Texture3DMipSamplingMode>(
                 parentElement, nodeBase, setNodesAsDirtyCallback, updateNodeViewsCallback,
                 "Mip Sampling Mode", "Change Mip Sampling Mode",
                 () => node.mipSamplingMode, (val) => node.mipSamplingMode = val);